requirements.txt
asgiref==3.8.1
django==5.0.6
django-debug-toolbar==4.4.2
django-js-asset==2.2.0
django-mptt==0.16.0
Pillow==10.3.0
pip==24.1.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1